#748 Refactor: ChangeDeckCard as organism
